+++ Runtime/GameFlow/IGameFlowController.cs
@@ -1,7 +1,10 @@
+using Jimothy.Systems.Core;
+
 namespace Jimothy.Systems.GameFlow
 {
     public interface IGameFlowController
     {
+        void Init();
         void StartGame();
         void EnterMainMenu();
         void RestartLevel();